// src/shared/api/generated/model/createMockRuleDtoMethod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Smart Mock Proxy
 * API сервиса Smart Mock Proxy: mock-ответы и проксирование на стенды
 * OpenAPI spec version: 0.1.0
 */

export type CreateMockRuleDtoMethod = typeof CreateMockRuleDtoMethod[keyof typeof CreateMockRuleDtoMethod];


export const CreateMockRuleDtoMethod = {
  GET: 'GET',
  POST: 'POST',
  PUT: 'PUT',
  PATCH: 'PATCH',
  DELETE: 'DELETE',
} as const;